refactor: Revise data fetching logic for separate location handling
